fix: #67-norminette-exept-exec/
