+++ Makefile
@@ -32,7 +32,9 @@ SRCS			=	src/main.c						\
 					src/parse/parse2.c				\
 					src/exec/exec.c					\
 					src/exec/is.c					\
+					src/exec/is2.c					\
 					src/exec/make_path.c			\
+					src/exec/make_path_2.c			\
 					src/redirect/redirect.c			\
 					src/redirect/heredoc.c			\
 					src/redirect/open_file.c		\
